+++ .gitlab-ci/container/container_post_build.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
 
-apt-get autoremove -y --purge
+if test -f /etc/debian_version; then
+    apt-get autoremove -y --purge
+fi
 
 # Clean up any build cache for rust.
 rm -rf /.cargo
 
-ccache --show-stats
+if test -x /usr/bin/ccache; then
+    ccache --show-stats
+fi
